remade province table to have an integer tax rate, added provinces to seed script, added a user tab to active admin

+++ db/migrate/20201130210144_create_provinces.rb
@@ -2,7 +2,7 @@ class CreateProvinces < ActiveRecord::Migration[6.0]
   def change
     create_table :provinces do |t|
       t.string :name
-      t.float :tax_rate
+      t.integer :tax_rate
 
       t.timestamps
     end
